+++ OptimoveCore/Sources/Classes/Constants/Endpoint.swift
@@ -5,9 +5,7 @@ import Foundation
 public struct Endpoints {
 
     public struct Logger {
-        public static let defaultEndpint = URL(string:
-            "https://us-central1-mobilepush-161510.cloudfunctions.net/reportLog"
-        )!
+        public static let defaultEndpint = URL(string: "https://mbaas.optimove.net/report/log")!
     }
 
     public struct Remote {
@@ -19,7 +17,7 @@ public struct Endpoints {
         public struct GlobalConfig {
             public static var url = URL(string: "https://sdk-cdn.optimove.net/configs/mobile")!
                 .appendingPathComponent("global")
-                .appendingPathComponent("v2")
+                .appendingPathComponent("v3")
                 .appendingPathComponent(SdkEnvironment.getBuildSetting(for: "OPTIMOVE_CONFIG_ENV_PATH", defaultValue: "prod"))
                 .appendingPathComponent("configs.json")
         }