+++ OptimoveSDK/Sources/Classes/Services/ServiceLocator.swift
@@ -29,6 +29,11 @@ final class ServiceLocator {
         )
     }()
 
+    /// Keeps as singleton in reason to share a session state between a service consumers.
+    private lazy var _locationManager: LocationService = {
+        return LocationServiceImpl()
+    }()
+
     private lazy var _deviceStateObserver: DeviceStateObserver = {
         return DeviceStateObserverFactory(
             statisticService: statisticService(),
@@ -146,7 +151,8 @@ final class ServiceLocator {
     func coreEventFactory() -> CoreEventFactory {
         return CoreEventFactoryImpl(
             storage: storage(),
-            dateTimeProvider: dateTimeProvider()
+            dateTimeProvider: dateTimeProvider(),
+            locationService: _locationManager
         )
     }